+++ Changelog.md
@@ -14,6 +14,7 @@
 - Added ETA fields to Shipping Methods
 - Added the Order::inChannel() scope (Foundation)
 - Added the Video module
+- Added the `HasVideos` trait to the `Product`, `MasterProduct`, `MasterProductVariant`, `Taxon` and `Taxonomy` models in the Foundation module
 
 ## 4.x Series